last bit of byte

+++ modules/client/src/main/kotlin/jp/co/soramitsu/iroha2/transaction/Instructions.kt
@@ -721,7 +721,7 @@ object Instructions {
     /**
      * Transfer an asset from the identifiable source.
      */
-    fun transferAsset(sourceId: AssetId, value: Long, destinationId: AssetId): Instruction {
+    fun transferAsset(sourceId: AssetId, value: Int, destinationId: AssetId): Instruction {
         return Instruction.Transfer(
             TransferBox(
                 sourceId = IdBox.AssetId(sourceId).evaluatesTo(),
